+++ CHANGELOG.md
@@ -17,6 +17,12 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added a new command, `proto shell`, that will initialize a list of tools into the environment and start an interactive shell session.
+  - By default, it will use the current shell, but you can override this with the `--shell` option.
+  - Example: `proto shell node@20 npm@10`
+
 #### ⚙️ Internal
 
 - Reworked identifier handling. This should be backwards compatible.